# specs/001-backend-api/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Cloud Native Todo Chatbot Backend API
  version: 1.0.0
  description: |
    RESTful API for todo management with chatbot intent integration.
    Stateless, cloud-native design for Kubernetes deployment.
  contact:
    name: Cloud Native Todo Team
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.com
    description: Production (placeholder)

paths:
  /api/v1/todos:
    post:
      summary: Create a new todo
      operationId: createTodo
      tags:
        - Todos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 500
                  description: Todo title; required and must be non-empty
              example:
                title: "Buy milk"
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Buy milk"
                completed: false
        '400':
          description: Invalid input (empty title, invalid JSON, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Title is required and must be non-empty"
                code: "INVALID_TITLE"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List all todos
      operationId: listTodos
      tags:
        - Todos
      responses:
        '200':
          description: List of all todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
              example:
                - id: "550e8400-e29b-41d4-a716-446655440000"
                  title: "Buy milk"
                  completed: false
                - id: "550e8400-e29b-41d4-a716-446655440001"
                  title: "Walk dog"
                  completed: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/todos/{id}:
    put:
      summary: Update todo completion status
      operationId: updateTodo
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Todo ID (UUID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - completed
              properties:
                completed:
                  type: boolean
                  description: Completion status
              example:
                completed: true
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Buy milk"
                completed: true
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Todo not found"
                code: "NOT_FOUND"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a todo
      operationId: deleteTodo
      tags:
        - Todos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Todo ID (UUID)
      responses:
        '204':
          description: Todo deleted successfully (no content returned)
        '404':
          description: Todo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/chat:
    post:
      summary: Process chatbot intent
      operationId: processChat
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  minLength: 1
                  description: Natural language message from chatbot
              example:
                message: "add todo buy milk"
      responses:
        '200':
          description: Intent parsed and processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              example:
                intent: "CREATE_TODO"
                status: "SUCCESS"
                result:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  title: "buy milk"
                  completed: false
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check (Kubernetes liveness probe)
      operationId: healthCheck
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: ["UP"]
              example:
                status: "UP"

  /ready:
    get:
      summary: Readiness check (Kubernetes readiness probe)
      operationId: readinessCheck
      tags:
        - Health
      responses:
        '200':
          description: Service is ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean
              example:
                ready: true
        '503':
          description: Service not ready (startup/shutdown)
          content:
            application/json:
              schema:
                type: object
                properties:
                  ready:
                    type: boolean
              example:
                ready: false

components:
  schemas:
    Todo:
      type: object
      required:
        - id
        - title
        - completed
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier (auto-generated)
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Todo title
        completed:
          type: boolean
          description: Completion status

    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_TITLE
            - INVALID_TYPE
            - INVALID_JSON
            - NOT_FOUND
            - INTERNAL_ERROR

    ChatResponse:
      type: object
      required:
        - intent
        - status
      properties:
        intent:
          type: string
          description: Parsed intent
          enum:
            - CREATE_TODO
            - LIST_TODOS
            - UPDATE_TODO
            - DELETE_TODO
            - UNKNOWN
        status:
          type: string
          description: Processing status
          enum:
            - SUCCESS
            - FAILED
        result:
          type: object
          description: Result object (if status=SUCCESS)
        error:
          type: string
          description: Error message (if status=FAILED)

tags:
  - name: Todos
    description: Todo CRUD operations
  - name: Chat
    description: Chatbot intent processing
  - name: Health
    description: Health and readiness probes